requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.1
feedparser==6.0.8
beautifulsoup4==4.10.0
supabase==2.13.0
openai==1.10.0
python-dateutil==2.8.2
pytz==2023.3
markdown==3.4.1
gunicorn==20.1.0
